Added ability to work with 'legacy' System.Configuration in Mono
All that really does is read machine.config from disk.

Adjusted configuration so the ImplicitMachineConfigHost is used along
with the original ConfigSystem. Added a check in IMCH to look for
and load an embedded machine.config if found.

Not all hostParams have the first array element as the filemap (e.g.
System.Web configuration) so use "as" instead of an explicit cast.